+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New `TransactionSigner` interface to allow for different signing agents
   in the `transactions` functions (e.g., makePreorder).
+- `putFile` can now optionally take the public key for which you want
+to encrypt the file. Thanks to @bodymindarts for this!
+
+### Changed
+- `encryptContent` now takes a public key instead of a private key to
+encrypt content for other users.
 
 ## [17.2.0]